refactor: 읽지 않은 신청 count 계산 기능 추가 (#78)

+++ api-module/src/main/java/org/devridge/api/domain/coffeechat/dto/response/GetAllCoffeeChatRequest.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class GetAllCoffeeChatRequest {
 
     private List<GetCoffeeChatRequestResponse> coffeeChatRequests;
-    private Long noReadCount;
+    private Integer noReadCount;
 }
